Gave SQS queues and SNS topics separate KMS keys
This way we can give roles/users decrypt/encrypt access just for those keys instead of for the root one of the account we use for the top security stuff.
