+++ tests/module_processors/test_aws_iam_role.py
@@ -169,4 +169,16 @@ class TestAwsIamRoleProcessor:
                 "Resource": ["${{module.topic.topic_arn}}"],
                 "Sid": "PublishSns",
             },
+            {
+                "Action": ["kms:GenerateDataKey", "kms:Decrypt"],
+                "Effect": "Allow",
+                "Resource": ["${{module.queue.kms_arn}}", "${{module.topic.kms_arn}}"],
+                "Sid": "KMSWrite",
+            },
+            {
+                "Action": ["kms:Decrypt"],
+                "Effect": "Allow",
+                "Resource": ["${{module.queue.kms_arn}}"],
+                "Sid": "KMSRead",
+            },
         ]
