+++ tests/module_processors/test_k8s_service.py
@@ -99,6 +99,21 @@ class TestK8sServiceProcessor:
                     "Resource": ["${{module.topic.topic_arn}}"],
                     "Sid": "PublishSns",
                 },
+                {
+                    "Action": ["kms:GenerateDataKey", "kms:Decrypt"],
+                    "Effect": "Allow",
+                    "Resource": [
+                        "${{module.queue.kms_arn}}",
+                        "${{module.topic.kms_arn}}",
+                    ],
+                    "Sid": "KMSWrite",
+                },
+                {
+                    "Action": ["kms:Decrypt"],
+                    "Effect": "Allow",
+                    "Resource": ["${{module.queue.kms_arn}}"],
+                    "Sid": "KMSRead",
+                },
             ],
         }
